requirements.txt
# Core Databricks and Spark
databricks-sdk>=0.35.0
pyspark>=3.5.0

# PDF Processing
PyMuPDF>=1.24.14
pypdfium2>=4.30.0

# Image Processing
Pillow>=10.4.0

# MLflow and LLM
mlflow[databricks]>=3.3.2

# Data Validation
pydantic>=2.10.3

# Utilities
python-dotenv>=1.0.1
tqdm>=4.67.1

# Testing
pytest>=8.3.4
pytest-mock>=3.14.0
